requirements.txt
absl-py==1.4.0
anyio==3.7.0
asgiref==3.6.0
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
Django==4.0.5
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.13.1
drf-yasg==1.21.7
fastapi==0.98.0
flatbuffers==23.5.26
fonttools==4.42.1
future==0.18.3
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httptools==0.5.0
huepy==1.2.1
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
inflection==0.5.1
iniconfig==2.0.0
instabot==0.117.0
joblib==1.3.2
keras==2.13.1rc0
kiwisolver==1.4.5
libclang==16.0.0
Markdown==3.4.3
markdown2==2.4.8
MarkupSafe==2.1.3
matplotlib==3.7.3
mock==5.1.0
moviepy==1.0.3
numpy==1.25.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pluggy==1.3.0
proglog==0.1.10
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.1.1
pypng==0.20220715.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.23.3
rsa==4.9
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
types-PyYAML==6.0.12.11
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
